Update regexp for memory allocation
The expression used assumed that the displayed line of the free memory starts with `Mem:`. This may NOT be happening in systems which are not in English. A workaround has been applied to grab only the line starting by `M` to fit outputs like:

```
$ free
              total       usado       libre  compartido búfer/caché  disponible
Memoria:    15656840     5287600     2484232     1300908     7885008     8787620
Swap:       1003516           0     1003516
```

Thus, the calculation can be performed:

```
$ LC_ALL=C free|awk '/^M[^:]+:/{print $2"/(2*1024)"}' | bc
7644
```